refactor: centralize optional-dependency typing shims
Move all *_INSTALLED feature flags and optional-dependency stub types
into advanced_alchemy/_typing.py as the single source of truth, and
re-export them through the public advanced_alchemy/typing.py.

- Add ATTRS_INSTALLED, CATTRS_INSTALLED, LITESTAR_INSTALLED, MSGSPEC_INSTALLED,
  NUMPY_INSTALLED, ORJSON_INSTALLED, PYDANTIC_INSTALLED to _typing.py
- Move BaseModel/Struct/TypeAdapter/AttrsInstance/cattrs/DTO/Empty stubs
  out of service/_typing.py into _typing.py
- Delete advanced_alchemy/service/_typing.py
- Re-route service/typing.py and other internal callers to import from
  advanced_alchemy.typing instead of reaching into private modules
