+++ advanced_alchemy/service/_util.py
@@ -16,10 +16,6 @@ from advanced_alchemy.exceptions import AdvancedAlchemyError
 from advanced_alchemy.filters import LimitOffset, StatementFilter
 from advanced_alchemy.service.pagination import OffsetPagination
 from advanced_alchemy.service.typing import (
-    ATTRS_INSTALLED,
-    CATTRS_INSTALLED,
-    MSGSPEC_INSTALLED,
-    PYDANTIC_INSTALLED,
     BaseModel,
     FilterTypeT,
     ModelDTOT,
@@ -30,6 +26,12 @@ from advanced_alchemy.service.typing import (
     is_attrs_schema,
     schema_dump,
     structure,
+)
+from advanced_alchemy.typing import (
+    ATTRS_INSTALLED,
+    CATTRS_INSTALLED,
+    MSGSPEC_INSTALLED,
+    PYDANTIC_INSTALLED,
 )
 
 if TYPE_CHECKING:
